refactor(tuner): page cell and ECU source loading out of their host files
